taxi24/requirements.txt
asgiref==3.3.1
astroid==2.5.1
coverage==5.5
Django==3.1.7
django-debug-toolbar==3.2
django-filter==2.4.0
django-restql==0.11.0
djangorestframework==3.12.4
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.7.2
pyPEG2==2.15.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
